wip: spid minor refactor response validation
